Move CrashServiceImpl to its own module
Part of a series of changes that aim to isolate each gRPC service into its own
module for better readability and maintenance. Currently, the Service module
contains multiple independent subsets of classes contributing to the different
services, so we extract those.

Test: Start OrbitService and Orbit UI in `--devmode`, verify OrbitService can be
crashed from the "Debug" menu.
