+++ CMakeLists.txt
@@ -161,6 +161,7 @@ add_subdirectory(src/ClientData)
 add_subdirectory(src/ClientModel)
 add_subdirectory(src/ClientProtos)
 add_subdirectory(src/ClientServices)
+add_subdirectory(src/CrashServiceImpl)
 add_subdirectory(src/DisplayFormats)
 add_subdirectory(src/FakeProducerSideService)
 add_subdirectory(src/FramePointerValidator)
