+++ src/Service/CMakeLists.txt
@@ -15,8 +15,6 @@ target_sources(ServiceLib PRIVATE
         CaptureServiceImpl.cpp
         CaptureServiceImpl.h
         CaptureStartStopListener.h
-        CrashServiceImpl.cpp
-        CrashServiceImpl.h
         FramePointerValidatorServiceImpl.cpp
         FramePointerValidatorServiceImpl.h
         MemoryInfoHandler.cpp
@@ -52,6 +50,7 @@ target_include_directories(ServiceLib PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 
 target_link_libraries(ServiceLib PUBLIC
         ApiLoader
+        CrashServiceImpl
         FramePointerValidator
         GrpcProtos
         Introspection
